feat(reports): extract useReports hook
- Extract all state, effects, and handlers to useReports hook (124 lines)
- Reports.jsx reduced from 256 to 131 lines (thin orchestrator)
- Keep existing ReportFilters and ReportViewer components
- Fix Reports.jsx import path for ReportFilters/ReportViewer
- All files ≤200 lines, zero behavior change
